+++ src/features/Dashboard/DashboardAsideLink.tsx
@@ -9,8 +9,8 @@ interface DashboardAsideLinkProps {
 
 export const DashboardAsideLink = ({ onClickCb, children, isSelected }: DashboardAsideLinkProps) => {
   const buttonCss = clsx(
-    "text-gray-600 dark:text-gray-400 hover:text-indigo-600 dark:hover:text-indigo-400 flex items-center gap-x-2.5 rounded-md px-2 py-1.5 text-sm transition hover:bg-gray-200 dark:hover:bg-gray-900  focus-visible:outline-2",
-    { "text-indigo-600 dark:text-indigo-400": isSelected }
+    "text-gray-300 flex items-center gap-x-2.5 rounded-md px-2 py-1.5 text-sm transition hover:bg-indigo-900 dark:hover:bg-indigo-700 focus-visible:outline-2",
+    { "text-white bg-indigo-900 dark:bg-indigo-700": isSelected }
   )
 
   return (